Added refresh on navigate to stock
solves the issue where if you navigate between stocks that have different graphs you have to refresh the page for the effect to occur (which most users wouldn't know to do, 1, and 2 would be annoying to ask the users to do).

Doesn't currently apply to the watchlist but it's commented in the proper location in case we see it happen there too. It seems that navigating from Dashboard to the stock/(inserttickerhere) page doesn't have the issue. Only stock to stock